Harden Stage A zero-shot evaluation (#310)
Close the audited Stage A evaluation gaps with fail-closed v2 contracts, shared baseline and identity closure, recomputed sealed-test selection, robustness gates, bounded common-draw bootstrap, atomic artifact publication, and exploit-focused regression coverage.
