+++ trade_rl/artifacts/atomic_write.py
@@ -1,0 +1,38 @@
+"""Small atomic file-write primitive shared by content-addressed artifacts."""
+
+from __future__ import annotations
+
+import os
+import uuid
+from pathlib import Path
+
+
+def _fsync_directory(path: Path) -> None:
+    if os.name == "nt":
+        return
+    descriptor = os.open(path, os.O_RDONLY)
+    try:
+        os.fsync(descriptor)
+    finally:
+        os.close(descriptor)
+
+
+def atomic_write_bytes(path: str | Path, payload: bytes) -> Path:
+    """Atomically replace *path* with fully flushed bytes and clean temporary files."""
+
+    output = Path(path)
+    output.parent.mkdir(parents=True, exist_ok=True)
+    temporary = output.with_name(f".{output.name}.tmp-{os.getpid()}-{uuid.uuid4().hex}")
+    try:
+        with temporary.open("xb") as handle:
+            handle.write(payload)
+            handle.flush()
+            os.fsync(handle.fileno())
+        os.replace(temporary, output)
+        _fsync_directory(output.parent)
+    finally:
+        temporary.unlink(missing_ok=True)
+    return output
+
+
+__all__ = ["atomic_write_bytes"]
